refact(Debtors): move destroy component file
